requirements.txt
inquirerpy
emoji
discord
python_dotenv
requests
